Fix DFS_traversal() Iterative
Set only popped nodes to DISCOVERED state.
Check if node is already discovered, if so POP and loop to while-loop
beginning.
We technically have not discovered a nodes adjacent nodes when we
iterate through the list. Hence only the parent node is DISCOVERED.
So we can still process edges of Discovered children.